Pull change added. Soft delete patient in Encounter Page
