+++ app/femr/common/IItemModelMapper.java
@@ -89,7 +89,6 @@ public interface IItemModelMapper {
      * @param weight             how much the patient weighs, may be null
      * @param pathToPatientPhoto filepath to the patient photo, may be null
      * @param photoId            id of the patients photo, may be null
-     * @param isDeleted          timestamp when a patient is deleted, may be null
      * @return a new PatientItem or null if processing fails, may be null
      */
     PatientItem createPatientItem(int id,
@@ -105,8 +104,7 @@ public interface IItemModelMapper {
                                   Integer heightInches,
                                   Float weight,
                                   String pathToPatientPhoto,
-                                  Integer photoId,
-                                  DateTime isDeleted);
+                                  Integer photoId);
 
     /**
      * Generate and provide an instance of PatientEncounterItem
