+++ app/femr/common/models/PatientItem.java
@@ -39,7 +39,6 @@ public class PatientItem {
     private Integer heightFeet;
     private Integer heightInches;
     private Float weight;
-    private DateTime isDeleted;
 
 
     public PatientItem(){
@@ -174,11 +173,5 @@ public class PatientItem {
 
     public void setFriendlyDateOfBirth(String friendlyDateOfBirth) {
         this.friendlyDateOfBirth = friendlyDateOfBirth;
-    }
-
-    public DateTime getIsDeleted(){return isDeleted;}
-
-    public void setIsDeleted(DateTime isDeleted){
-        this.isDeleted = isDeleted;
     }
 }
